% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBase.R
\name{addBase}
\alias{addBase}
\title{addBase()}
\usage{
addBase(myGIO, base)
}
\arguments{
\item{base}{a string indicating the base map to be used}

\item{myGIo}{the widget object}
}
\value{
myGIO object with base layer attached
}
\description{
Adds the base layer map
}
