% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLayer.R
\name{addLayer}
\alias{addLayer}
\title{addLayer()}
\usage{
addLayer(myGIO, type, label, data = NULL, mapping = NULL,
  options = NULL)
}
\arguments{
\item{type}{a string indicating the layer to be added ("world", "us")}

\item{label}{a string unique to the layer for the chart}

\item{data}{is a dataframe to be mapped tot he base layer}

\item{mapping}{is a list object mapping the data to the map features; requires a data key and map key}

\item{options}{is a vector of objects to set and assign layer behaviors}

\item{myGIo}{the widget object}
}
\value{
myGIO object with base layer attached
}
\description{
Adds a layer to the map
}
