% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPolygonFill.R
\name{assignPolygonFill}
\alias{assignPolygonFill}
\title{assignPolygonFill()}
\usage{
assignPolygonFill(ids, propertyId, color = "steelblue")
}
\arguments{
\item{ids}{a vector of polygon ids to fill}

\item{propertyId}{is a string indicating the polygon property for matching the id}

\item{color}{is a string with any HTML/CSS recognized color}
}
\value{
list object with layer options
}
\description{
Assigns color to a vector of polygons as layer option
}
