% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGIO.R
\name{myGIO}
\alias{myGIO}
\title{<Add Title>}
\usage{
myGIO(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a data frame}

\item{width}{a string indicating pixels or percent width}

\item{height}{a string indicating pixels height}

\item{elementId}{a unique htmlwidgets identifier}
}
\value{
myGIO object
}
\description{
<Add Description>
}
