% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGeoJSON.R
\name{readGeoJSON}
\alias{readGeoJSON}
\title{readGeoJSON()}
\usage{
readGeoJSON(myGIO, file_path)
}
\arguments{
\item{file_path}{}

\item{myGIo}{the widget object}
}
\value{
myGIO object with base layer attached
}
\description{
reads a geoJSON file from a specific path
}
