% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPolygonZoom.R
\name{setPolygonZoom}
\alias{setPolygonZoom}
\title{setPolytonZoom()}
\usage{
setPolygonZoom(behavior = "center", zoomScale = 2)
}
\arguments{
\item{behavior}{a string indicating the desired zoom behavior for the layer: "center"}

\item{zoomScale}{is a numeric value setting the max value for the zooms scale}
}
\value{
list object with layer options
}
\description{
Sets the zoom behavior of polygons as layer option
}
